refactor(cockpit): extract review evidence helpers
